move auth_middleware to /middleware/auth_middleware; add getHeaderTypeAndValue
